Introduce the persistent store and move the in-mem facility only to tests
