+++ WebDev/src/main/java/com/webdev/Application.java
@@ -15,5 +15,4 @@ public class Application {
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
-
 }
