+++ WebDev/src/main/resources/db/schema.sql
@@ -17,44 +17,42 @@
  * 
  */
 create table MenuItem (
-	id bigint identity(1,1) not null,
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	item_name varchar(50) not null, 
 	menu_type varchar(15) not null,-- Indian, Italian, Continental,.. others
 	food_type varchar(15) not null,-- Food, Drink, Condiments,.. others
 	price numeric(19,2) not null,
-	inactive varchar(1) not null,
-	Primary Key (id)
+	inactive varchar(1) not null--,
+	--Primary Key (id)
 );
 
 create table Orders (
-	id bigint identity(1,1) not null, 
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	placement_date date not null, 
 	delivery_date date not null,
 	status varchar(1) not null, -- O for Open, P for Processing, D for Delivered
 	total_value numeric(19,2) not null,
-	user_id varchar(20) null,  --//create an index for this column
-	Primary Key (id)
+	user_id varchar(20) null  --//create an index for this column
+	--Primary Key (id)
 );
 
 create table Order_Item (
-	id bigint identity(1,1) not null,
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	order_id int not null, -- FK to Order.id
 	item_id int not null , -- FK to MenuItem.id
 	quantity int not null, 
 	price numeric(19,2) not null, 
-	notes varchar(250),
-	Primary Key (id)
+	notes varchar(250)--,	Primary Key (id)
 );
 
-create table User (
-	id bigint identity(1,1) not null,
+create table Users (
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	user_id varchar(250) not null unique,
 	password varchar(250) not null,
 	first_name varchar(250) not null,
 	last_name varchar(250) not null,
 	gender char(1) not null,
-	enabled TINYINT NOT NULL DEFAULT 1 ,
-	Primary Key (id)
+	enabled boolean NOT NULL DEFAULT true --,Primary Key (id)
 );
 
 create table role(
@@ -63,16 +61,15 @@ create table role(
 );
 
 create table user_role (
-	id bigint identity(1,1) not null, 
+	id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
 	user_id varchar(250) not null, 
-	role varchar(250) not null, 
-	primary key (id),
-	unique key (user_id, role)
+	role varchar(250) not null,--, 	primary key (id),
+	unique (user_id, role)
 );
 
 Alter table Orders
 Add constraint order_User_FK
-Foreign Key (user_id) References User(user_id);
+Foreign Key (user_id) References Users(user_id);
 
 Alter table Order_Item
 Add constraint orderItem_Order_FK
@@ -84,7 +81,7 @@ Foreign Key (item_id) References MenuItem;
 
 Alter table user_role
 Add Constraint user_role_user_FK
-Foreign Key (user_id) References User(user_id);
+Foreign Key (user_id) References Users(user_id);
 
 Alter table user_role
 Add Constraint user_role_role_FK
