+++ WebDev/test/main/java/com/webdev/controllers/LoginControllerTest.java
@@ -11,6 +11,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
@@ -19,10 +21,14 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 import com.webdev.Application;
+import com.webdev.config.DBConfiguration;
+import com.webdev.config.MvcConfiguration;
+import com.webdev.config.TestBaseConfig;
+import com.webdev.data.util.hibernate.HibernateConfig;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@WebAppConfiguration
-@ContextConfiguration(classes={Application.class})
+//@WebAppConfiguration
+@SpringBootTest
 public class LoginControllerTest {
 	
     @Autowired
