+++ WebDev/test/main/resources/test.properties
@@ -1,0 +1,4 @@
+# To change this license header, choose License Headers in Project Properties.
+# To change this template file, choose Tools | Templates
+# and open the template in the editor.
+spring.jpa.hibernate.ddl-auto=
